Include hiredis sources directly in package
The other option is to use a configure script which seems better,
but requires autoconf/autotools, separate windows installation,
etc.  Provided this works OK on Windows, I think that the all-in
approach is probably simpler, and possibly more robust.

The update_hiredis script will, given a path to the hiredis
tarball, unpack the required files and put everything into place.